move theme provider
