+++ src/assets/styles/index.css
@@ -1,0 +1,6 @@
+html,
+body,
+#root {
+  min-height: 100%;
+  height: 100%;
+}
