+++ src/utils/theme/darkPalette.js
@@ -1,0 +1,15 @@
+const darkPalette = {
+  primary: {
+    main: "#5964E0",
+    dark: "#939BF4",
+  },
+  secondary: {
+    main: "rgba(255, 255, 255, .1)",
+    dark: "rgba(255, 255, 255, .35)",
+    contrastText: "#FFF",
+  },
+  background: {
+    default: "#19202D",
+  },
+};
+export default darkPalette;
