+++ src/utils/theme/index.js
@@ -1,0 +1,6 @@
+// Themes
+export { default as darkPalette } from "./darkPalette";
+export { default as lightPalette } from "./lightPalette";
+
+// Overrides
+export { default as stylesOverrides } from "./stylesOverrides";
