+++ src/utils/theme/lightPalette.js
@@ -1,0 +1,15 @@
+const lightPalette = {
+  primary: {
+    main: "#5964E0",
+    dark: "#939BF4",
+  },
+  secondary: {
+    main: "rgba(89, 100, 224, 0.1)",
+    dark: "rgba(89, 100, 224, 0.35)",
+    contrastText: "#5964E0",
+  },
+  background: {
+    default: "#F4F6F8",
+  },
+};
+export default lightPalette;
